test(sandbox): isolate POSIX process-group regression
