+++ sdk/tools/sandbox/sandbox_process_group_unix_test.go
@@ -1,0 +1,77 @@
+//go:build aix || android || darwin || dragonfly || freebsd || illumos || linux || netbsd || openbsd || solaris
+
+package sandbox
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"syscall"
+	"testing"
+	"time"
+
+	"github.com/timwhitez/agent-sdk-golang/sdk/tools"
+)
+
+func TestBashTool_KillsProcessGroupOnTimeout(t *testing.T) {
+	root := t.TempDir()
+	s, err := New(root)
+	if err != nil {
+		t.Fatalf("new: %v", err)
+	}
+	deps := tools.NewContainer()
+	tools.Provide(deps, Key, func(context.Context) (*Sandbox, error) { return s, nil })
+	tools.Provide(deps, ConfirmKey, func(context.Context) (Confirmer, error) { return allowConfirmer{}, nil })
+
+	ctx := tools.WithToolResultMetadata(context.Background())
+	command := `sleep 30 & child=$!; echo $child > child.pid; wait $child`
+	out, err := bashTool().Execute(ctx, fmt.Sprintf(`{"command":%q,"timeout":1}`, command), deps)
+	if !errors.Is(err, context.DeadlineExceeded) {
+		t.Fatalf("expected deadline exceeded, got %v", err)
+	}
+	if !strings.Contains(strings.ToLower(out.PlainText()), "timed out") {
+		t.Fatalf("expected timed out output, got %q", out.PlainText())
+	}
+
+	pidPath := filepath.Join(root, "child.pid")
+	var pid int
+	deadline := time.Now().Add(2 * time.Second)
+	for time.Now().Before(deadline) {
+		data, readErr := os.ReadFile(pidPath)
+		if readErr == nil {
+			pid, err = strconv.Atoi(strings.TrimSpace(string(data)))
+			if err != nil {
+				t.Fatalf("parse child pid: %v", err)
+			}
+			break
+		}
+		time.Sleep(20 * time.Millisecond)
+	}
+	if pid <= 0 {
+		t.Fatalf("expected child pid to be written at %s", pidPath)
+	}
+	t.Cleanup(func() {
+		_ = syscall.Kill(pid, syscall.SIGKILL)
+	})
+
+	aliveDeadline := time.Now().Add(2 * time.Second)
+	for time.Now().Before(aliveDeadline) {
+		if !processExists(pid) {
+			return
+		}
+		time.Sleep(25 * time.Millisecond)
+	}
+	t.Fatalf("expected child process %d to be terminated with parent timeout", pid)
+}
+
+func processExists(pid int) bool {
+	if pid <= 0 {
+		return false
+	}
+	err := syscall.Kill(pid, 0)
+	return err == nil || errors.Is(err, syscall.EPERM)
+}
